+++ pkg/networkpath/traceroute/traceroute.go
@@ -8,14 +8,9 @@ package traceroute
 
 import (
 	"context"
-	"fmt"
-	"net/http"
 	"time"
 
-	sysprobeclient "github.com/DataDog/datadog-agent/cmd/system-probe/api/client"
-	sysconfig "github.com/DataDog/datadog-agent/cmd/system-probe/config"
 	"github.com/DataDog/datadog-agent/pkg/networkpath/payload"
-	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
 type (
@@ -48,40 +43,3 @@ type (
 		Run(context.Context) (payload.NetworkPath, error)
 	}
 )
-
-func getTraceroute(client *http.Client, clientID string, host string, port uint16, protocol payload.Protocol, maxTTL uint8, timeout time.Duration) ([]byte, error) {
-	httpTimeout := timeout*time.Duration(maxTTL) + 10*time.Second // allow extra time for the system probe communication overhead, calculate full timeout for TCP traceroute
-	log.Tracef("Network Path traceroute HTTP request timeout: %s", httpTimeout)
-	ctx, cancel := context.WithTimeout(context.Background(), httpTimeout)
-	defer cancel()
-
-	url := sysprobeclient.ModuleURL(sysconfig.TracerouteModule, "/traceroute/"+host) + fmt.Sprintf("?client_id=%s&port=%d&max_ttl=%d&timeout=%d&protocol=%s", clientID, port, maxTTL, timeout, protocol)
-	req, err := http.NewRequestWithContext(ctx, "GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("Accept", "application/json")
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode == http.StatusBadRequest {
-		body, err := sysprobeclient.ReadAllResponseBody(resp)
-		if err != nil {
-			return nil, fmt.Errorf("traceroute request failed: url: %s, status code: %d", req.URL, resp.StatusCode)
-		}
-		return nil, fmt.Errorf("traceroute request failed: url: %s, status code: %d, error: %s", req.URL, resp.StatusCode, string(body))
-	} else if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("traceroute request failed: url: %s, status code: %d", req.URL, resp.StatusCode)
-	}
-
-	body, err := sysprobeclient.ReadAllResponseBody(resp)
-	if err != nil {
-		return nil, err
-	}
-
-	return body, nil
-}
